+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log (vscode-proxy)
 
+## 0.18.0 (February 11th, 2019; FINAL VERSION => DEPRECATED)
+
+* **the extension is now marked as DEPRECATED ... it is RECOMMENDED to use [vscode-powertools](https://marketplace.visualstudio.com/items?itemName=ego-digital.vscode-powertools) by [e.GO Digital](https://github.com/egodigital)**
+* if you have suggestions and other kind of issues for that new extension, feel free to [open an issue here](https://github.com/egodigital/vscode-powertools/issues)
+
 ## 0.17.1 (October 14th, 2017; multi root support)
 
 * started to refactor to new, upcoming [Multi Root Workspace API](https://github.com/Microsoft/vscode/wiki/Extension-Authoring:-Adopting-Multi-Root-Workspace-APIs)